+++ src/dungeon/func_80285E80.c
@@ -45,7 +45,7 @@ extern u8 D_800E39C8[12];
 
 extern void func_800194C4(s16 arg0);
 extern void func_80019648(void);
-extern s32 func_80019684(s16 arg0, s32 arg1, s32 arg2);
+extern s32 func_80019684(s16 arg0, s32 arg1, s16 arg2);
 extern void func_80019A74(void);
 extern s32 func_80033BC0(s32 arg0);
 extern void func_80044698(void);
