+++ src/dungeon/func_81941338.c
@@ -33,7 +33,7 @@ extern s32 func_800243D8(void *, void *, void *);
 extern s32 func_800B8FC8(void *, void *, void *, s32, s32);
 extern s32 func_8003DF74(void *, void *, void *, s32);
 extern s32 func_80069EF8(void);
-extern s32 func_80024798(void *, s32, s32, s32, s32, s32, s32);
+extern s32 func_80024798(void *, s32, s32, s32, s16, s16, s16);
 extern void *func_8003FC64(s32);
 extern s32 func_8004491C(void *, void *);
 extern s32 func_80053EF0(s32);
